+++ affinidi-did-resolver-cache-sdk/src/networking/request_queue.rs
@@ -142,7 +142,7 @@ mod tests {
     use std::collections::HashMap;
 
     use blake2::{Blake2s256, Digest};
-    use rand::{distributions::Alphanumeric, Rng};
+    use rand::{distr::Alphanumeric, Rng};
     use tokio::sync::oneshot::{self, Sender};
 
     use crate::{
@@ -290,7 +290,7 @@ mod tests {
     }
 
     fn _unique_id() -> String {
-        rand::thread_rng()
+        rand::rng()
             .sample_iter(&Alphanumeric)
             .take(8)
             .map(char::from)