+++ affinidi-did-resolver-cache-server/src/common.rs
@@ -1,4 +1,4 @@
-use rand::{distributions::Alphanumeric, Rng};
+use rand::{distr::Alphanumeric, Rng};
 use serde::{de::DeserializeOwned, Serialize};
 
 /// Helps with deserializing the generic data field in the SuccessResponse struct
@@ -6,7 +6,7 @@ pub trait GenericDataStruct: DeserializeOwned + Serialize {}
 
 // Creates a random transaction identifier for each transaction
 pub(crate) fn create_session_id() -> String {
-    rand::thread_rng()
+    rand::rng()
         .sample_iter(&Alphanumeric)
         .take(8)
         .map(char::from)